Update TutorialGui.py
Change the GUI looks a bit and attempted to make a function where it gets the password of the username and compares it to the password inputted. If it does then it says a message. There is no function which expands on this as it's just a log in system. The function doing this is GettingPasswordFromDatabaseAndChecking().
Also made it so that the window is destroyed after the login or register button is pressed. 
Things to do:
Make the GettingPasswordFromDatabaseAndChecking work properly as it's not fetching the correct data
Maybe use recursission to force the user to input valid data. So if it's not valid or incorrect then current window is destroyed and the function registerpage() is called again.
Change the code so that only window is visible at one time by destroying a previous one before a new one can be made. Maybe even implement q button at the top left to go back.